requirements.txt
openai>=1.12.0
pyautogui>=0.9.54
pillow>=10.0.0
python-dotenv>=1.0.0
